requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
openai==1.50.0
httpx==0.27.0
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.2
